General code optimization to reduce number of warnings triggered, and to remove additional unneeded LotR components

+++ gemp-lotr/gemp-lotr-server/src/test/java/com/gempukku/lotro/merchant/ParametrizedMerchantTest.java
@@ -25,11 +25,11 @@ public class ParametrizedMerchantTest extends AbstractAtTest {
 
         Date currentTime = new Date(1000 * 60 * 60 * 24 * 5L);
 
-        assertEqualsMoreOrLess(1100, (int) merchant.getCardSellPrice("1_1", firstTrans));
-        assertEqualsMoreOrLess(900, (int) merchant.getCardBuyPrice("1_2", firstTrans));
+        assertEqualsMoreOrLess(1100, merchant.getCardSellPrice("1_1", firstTrans));
+        assertEqualsMoreOrLess(900, merchant.getCardBuyPrice("1_2", firstTrans));
 
-        assertEqualsMoreOrLess(1000, (int) merchant.getCardSellPrice("1_1", currentTime));
-        assertEqualsMoreOrLess(1000, (int) merchant.getCardBuyPrice("1_2", currentTime));
+        assertEqualsMoreOrLess(1000, merchant.getCardSellPrice("1_1", currentTime));
+        assertEqualsMoreOrLess(1000, merchant.getCardBuyPrice("1_2", currentTime));
     }
 
     //    @Test
@@ -50,8 +50,8 @@ public class ParametrizedMerchantTest extends AbstractAtTest {
 
         System.out.println("-2,1000,700,1000,700");
         for (long time = 0; time < hour * 24 * 35L; time += hour * 2) {
-            System.out.println(time / hour + "," + ((int) merchant.getCardSellPrice("1_1", new Date(time))) + "," + ((int) merchant.getCardBuyPrice("1_1", new Date(time)))
-                    + "," + ((int) merchant.getCardSellPrice("1_2", new Date(time))) + "," + ((int) merchant.getCardBuyPrice("1_2", new Date(time))));
+            System.out.println(time / hour + "," + merchant.getCardSellPrice("1_1", new Date(time)) + "," + merchant.getCardBuyPrice("1_1", new Date(time))
+                    + "," + merchant.getCardSellPrice("1_2", new Date(time)) + "," + merchant.getCardBuyPrice("1_2", new Date(time)));
         }
     }
 
@@ -69,7 +69,7 @@ public class ParametrizedMerchantTest extends AbstractAtTest {
         long hour = 1000 * 60 * 60;
 
         for (long time = 0; time < hour * 24 * 35L; time += hour * 2)
-            System.out.println(time / hour + "," + ((int) merchant.getCardSellPrice("1_1", new Date(time))) + "," + ((int) merchant.getCardBuyPrice("1_1", new Date(time))));
+            System.out.println(time / hour + "," + merchant.getCardSellPrice("1_1", new Date(time)) + "," + merchant.getCardBuyPrice("1_1", new Date(time)));
     }
 
     private void assertEqualsMoreOrLess(int expected, int given) {
